feat: first pass on transaction workflow

+++ src/main/java/com/jacobhampton/techtest/account/dto/AccountResponseDto.java
@@ -9,7 +9,7 @@ public record AccountResponseDto(
         String sortCode,
         String name,
         AccountType accountType,
-        int balance,
+        double balance,
         Currency currency,
         String createdTimestamp,
         String updatedTimestamp
